fix commmmit

+++ app/database/migrations/2014_09_25_104219_create_productos_table.php
@@ -17,8 +17,8 @@ class CreateProductosTable extends Migration {
 			$table->string('codigo', 50);
 			$table->string('descripcion',100);
 			$table->integer('existencia')->default(0);
-			$table->integer('p_costo');
-			$table->decimal('p_publico', 8, 2);
+			$table->decimal('p_costo', 8, 5);
+			$table->decimal('p_publico', 8, 5);
 			$table->integer('stock_minimo')->default(0);
 			$table->integer('inactivo');
